Removed unused libraries  Changed output format of information about the application changed the name of the function
